+++ enguino/printEthernet.h
@@ -5,17 +5,17 @@
 // Outputs to 'client', remember to 'flush()' the buffer after last print
 
 //  This file is part of Enguino.
-//  
+//
 //  Enguino is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
 //  the Free Software Foundation, either version 3 of the License, or
 //  (at your option) any later version.
-//  
+//
 //  Enguino is distributed in the hope that it will be useful,
 //  but WITHOUT ANY WARRANTY; without even the implied warranty of
 //  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 //  GNU General Public License for more details.
-//  
+//
 //  You should have received a copy of the GNU General Public License
 //  along with Enguino.  If not, see <http://www.gnu.org/licenses/>.
 
@@ -24,7 +24,7 @@
 
 
 char buffer[64];
-int bfree = 64;
+byte bfree = 64;
 char *bufp = buffer;
 
 void flush() {
@@ -33,9 +33,9 @@ void flush() {
   bufp = buffer;
 }
 
-void print(string cp, int len) {
+void print(string cp, short len) {
   while (len) {
-    int n = len;
+    short n = len;
     if (n > bfree)
       n = bfree;
     memcpy(bufp, cp, n);
@@ -44,14 +44,14 @@ void print(string cp, int len) {
     if (bfree == 0)
       flush();
     cp += n;
-    len -= n;     
+    len -= n;
   }
 }
 
-void print_P(const __FlashStringHelper * cp, int len) {
+void print_P(const __FlashStringHelper * cp, short len) {
   PGM_P pp = reinterpret_cast<PGM_P>(cp);
   while (len) {
-    int n = len;
+    short n = len;
     if (n > bfree)
       n = bfree;
     memcpy_P(bufp, pp, n);
@@ -60,7 +60,7 @@ void print_P(const __FlashStringHelper * cp, int len) {
     if (bfree == 0)
       flush();
     pp += n;
-    len -= n;     
+    len -= n;
   }
 }
 
@@ -77,7 +77,7 @@ void print(char c) {
   print(&c,1);
 }
 
-void print(int n, int decimal) {
+void print(int n, short decimal) {
   char buf[7];
   buf[6] = 0;
   char *cp = buf + 6;
@@ -86,7 +86,7 @@ void print(int n, int decimal) {
       print("inop");
       return;
     }
-       
+
     print('-');
     n = -n;
   }
@@ -112,7 +112,7 @@ void print_n_close(int n) {
   print_n_close();
 }
 
-void print_g_translate(int x, int y) {
+void print_g_translate(short x, short y) {
   print_P(F("<g transform='translate("));
   print(x);
   print(' ');
@@ -132,4 +132,3 @@ void print_text_close(string cp) {
   print(cp);
   print_text_close();
 }
-
