+++ enguino/tcTemp.h
@@ -5,17 +5,17 @@
 // also handles periodic real-time activites for Enguino
 //
 //  This file is part of Enguino.
-//  
+//
 //  Enguino is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
 //  the Free Software Foundation, either version 3 of the License, or
 //  (at your option) any later version.
-//  
+//
 //  Enguino is distributed in the hope that it will be useful,
 //  but WITHOUT ANY WARRANTY; without even the implied warranty of
 //  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 //  GNU General Public License for more details.
-//  
+//
 //  You should have received a copy of the GNU General Public License
 //  along with Enguino.  If not, see <http://www.gnu.org/licenses/>.
 
@@ -30,12 +30,12 @@
 #define PINSC 13 //TC ADC Serial Clock (SCK)
 #define PINCS 9  //TC ADC Chip Select
 
-volatile int tcTemp[9];    // in quarter deg. C, tcTemp[8] is the interal reference temp, disable IRQ's to access these
+volatile short tcTemp[9];    // in quarter deg. C, tcTemp[8] is the interal reference temp, disable IRQ's to access these
 
 volatile bool eighthSecondTick;
 volatile byte eighthSecondCount = 0;
 
-int readSPI() {
+short readSPI() {
   word v = 0;
   for (byte i=16; i; i--) {
     v <<= 1;
@@ -49,14 +49,14 @@ int readSPI() {
 }
 
 void tcTempSetup() {
-  pinMode(PINEN, OUTPUT);     
-  pinMode(PINA0, OUTPUT);    
-  pinMode(PINA1, OUTPUT);    
-  pinMode(PINA2, OUTPUT);    
-  pinMode(PINSO, INPUT);    
-  pinMode(PINCS, OUTPUT);    
-  pinMode(PINSC, OUTPUT);    
-  
+  pinMode(PINEN, OUTPUT);
+  pinMode(PINA0, OUTPUT);
+  pinMode(PINA1, OUTPUT);
+  pinMode(PINA2, OUTPUT);
+  pinMode(PINSO, INPUT);
+  pinMode(PINCS, OUTPUT);
+  pinMode(PINSC, OUTPUT);
+
   digitalWrite(PINEN, HIGH);   // enable the mux all the time
   digitalWrite(PINSC, LOW);    // put clock in low
   digitalWrite(PINCS, LOW);    // stop conversion, start serial interface
@@ -64,11 +64,11 @@ void tcTempSetup() {
   // Timer0's overflow is used for millis() - setup to interrupt
   // in the middle and call the "Compare A" function below
   OCR0A = 0x80;
-  TIMSK0 |= _BV(OCIE0A);  
+  TIMSK0 |= _BV(OCIE0A);
 }
 
 // Interrupt is called every millisecond (a little slower, actually 976.5625 Hz)
-SIGNAL(TIMER0_COMPA_vect) 
+SIGNAL(TIMER0_COMPA_vect)
 {
   static byte ms = 0;
   static byte ch = 0;
@@ -76,13 +76,13 @@ SIGNAL(TIMER0_COMPA_vect)
 #if DEBUG
   checkFreeRam();
 #endif
-  
+
   if (ms == 0) {
     // select the thermocouple channel on the mux
-    digitalWrite(PINA0, ch&1); 
-    digitalWrite(PINA1, ch&2); 
+    digitalWrite(PINA0, ch&1);
+    digitalWrite(PINA1, ch&2);
     digitalWrite(PINA2, ch&4);
-    // ... wait a while for the capacitor on the ADC input to charge (< .1 mS actually needed)     
+    // ... wait a while for the capacitor on the ADC input to charge (< .1 mS actually needed)
   }
   else if (ms == 21) {
     // begin conversion
@@ -93,13 +93,13 @@ SIGNAL(TIMER0_COMPA_vect)
   }
   else if (ms == 121) {   // spec says 100mS, IRQ's are a bit slower, so this is >100mS
     // stop conversion, start serial interface
-    digitalWrite(PINCS, LOW); 
-    // 100nS min. delay implied 
-    
-    int rawTC = readSPI();
-    int rawIT = readSPI();
-    
-    int tempC = rawTC / 4;
+    digitalWrite(PINCS, LOW);
+    // 100nS min. delay implied
+
+    short rawTC = readSPI();
+    short rawIT = readSPI();
+
+    short tempC = rawTC / 4;
     if (rawTC & 1) {
       if (rawIT & 7)
         tempC = FAULT;
@@ -110,8 +110,8 @@ SIGNAL(TIMER0_COMPA_vect)
       tcTemp[8] = rawIT / 64; // internal temperature reduced to quarter degree C
       ch = 0;
     }
-    ms = 255; // ++ will make this 0 
-  
+    ms = 255; // ++ will make this 0
+
     eighthSecondCount++;
     if ((eighthSecondCount&3) == 0) // every half second
       updateFuelFlow();
@@ -120,4 +120,3 @@ SIGNAL(TIMER0_COMPA_vect)
   }
   ms++;
 }
-
